Props2 migration followup.

+++ tools/auralink/auraexport_csv.py
@@ -125,9 +125,9 @@ if args.flight:
             last_counter = counter
             if not located:
                 if gps_node.getInt('satellites') >= 5:
-                    lat = gps_node.getFloat('latitude_deg')
-                    lon = gps_node.getFloat('longitude_deg')
-                    sec = gps_node.getFloat('unix_time_sec')
+                    lat = gps_node.getDouble('latitude_deg')
+                    lon = gps_node.getDouble('longitude_deg')
+                    sec = gps_node.getDouble('unix_time_sec')
                     located = True
             current.compute_derived_data()
             category = logical_category(id)
@@ -173,9 +173,9 @@ for key in sorted(data):
         
 print()
 print("Total log time: %.1f min" % (total_time / 60.0))
-print("Flight timer: %.1f min" % (status_node.getFloat('flight_timer') / 60.0))
-print("Autopilot time: %.1f min" % (status_node.getFloat('local_autopilot_timer') / 60.0))
-print("Distance flown: %.2f nm (%.2f km)" % (status_node.getFloat('flight_odometer')*m2nm, status_node.getFloat('flight_odometer')*0.001))
+print("Flight timer: %.1f min" % (status_node.getDouble('flight_timer') / 60.0))
+print("Autopilot time: %.1f min" % (status_node.getDouble('local_autopilot_timer') / 60.0))
+print("Distance flown: %.2f nm (%.2f km)" % (status_node.getDouble('flight_odometer')*m2nm, status_node.getDouble('flight_odometer')*0.001))
 print("Battery Usage: %.0f mah" % apm2_node.getInt("extern_current_mah"))
 print()
